Create uploading-with-php.php
This is a simple way to upload an image using PHP. Of note, this is not the most secure way to have a user upload an image. 
For this to be more secure, a switch statement can be used to validate the file type being uploaded with pre-defined filetypes that 
you specify, and also create a pre-determined filename for the uploaded image.